requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
supervision>=0.18.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
onnx>=1.14.0
onnxruntime>=1.15.0
streamlit>=1.28.0
PyYAML>=6.0
